SCRS-7210 Test setup using S4L builder
